# Conflicts:
#	.gitignore
#	README.md
#	activity/src/main/dml/activity.dml
#	activity/src/main/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityProvider.java
#	activity/src/main/java/pt/ulisboa/tecnico/softeng/activity/domain/Booking.java
#	activity/src/main/java/pt/ulisboa/tecnico/softeng/activity/services/local/ActivityInterface.java
#	bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Account.java
#	bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Bank.java
#	bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/services/local/dataobjects/BankOperationData.java
#	bank/src/test/java/pt/ulisboa/tecnico/softeng/bank/services/local/BankInterfaceProcessPaymentMethodTest.java
#	broker/src/main/dml/broker.dml
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/BookRoomState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/Broker.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/ProcessPaymentState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/ReserveActivityState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/services/local/dataobjects/BulkData.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/services/remote/ActivityInterface.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/services/remote/BankInterface.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/services/remote/HotelInterface.java
#	broker/src/main/resources/templates/adventures.html
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/AdventureSequenceTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/BookRoomStateMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/BulkRoomBookingProcessBookingMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/ProcessPaymentStateProcessMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/ReserveActivityStateProcessMethodTest.java
#	hotel/src/main/dml/hotel.dml
#	hotel/src/main/java/pt/ulisboa/tecnico/softeng/hotel/domain/Hotel.java
#	hotel/src/main/java/pt/ulisboa/tecnico/softeng/hotel/services/local/HotelInterface.java
#	pom.xml

+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/services/local/dataobjects/BankOperationData.java
@@ -8,7 +8,7 @@ public class BankOperationData {
 	private String reference;
 	private String type;
 	private String iban;
-	private double value;
+	private Double value;
 	private DateTime time;
 
 	public BankOperationData() {
@@ -50,7 +50,7 @@ public class BankOperationData {
 		return this.value;
 	}
 
-	public void setValue(int value) {
+	public void setValue(double value) {
 		this.value = value;
 	}
